purchases add

+++ src/controllers/cart.controllers.js
@@ -53,17 +53,20 @@ const getOne = catchError(async(req, res) => {
 });
 
 const remove = catchError(async(req, res) => {
+    const userId = req.user.id
     const { id } = req.params;
-    const result = await Cart.destroy({ where: {id} });
+    const result = await Cart.destroy({ where: {id, userId} });
     if(!result) return res.sendStatus(404);
     return res.sendStatus(204);
 });
 
 const update = catchError(async(req, res) => {
+    const userId = req.user.id
     const { id } = req.params;
+    const { quantity } = req.body
     const result = await Cart.update(
-        req.body,
-        { where: {id}, returning: true }
+        { quantity } ,
+        { where: {id, userId}, returning: true }
     );
     if(result[0] === 0) return res.sendStatus(404);
     return res.json(result[1][0]);
